Separate Work component from page route component
